requirements/local.txt
alembic==1.13.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.6
crypto==1.4.1
distlib==0.3.8
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.1
filelock==3.13.1
h11==0.14.0
identify==2.5.34
idna==3.4
ifaddr==0.2.0
jose==1.0.0
Mako==1.3.1
MarkupSafe==2.1.4
Naked==0.1.32
nodeenv==1.8.0
passlib==1.7.4
platformdirs==4.2.0
pre-commit==3.6.1
psycopg2==2.9.7
py_eureka_client==0.11.9
pyasn1==0.5.0
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.40.5
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.25.0
